fixed a small bug within lib/screencomponents/animexpanded
the bug would've allowed a user to break the rotation of a button if they pressed it before it was completed with the animation, this would've flipped it upside down & been unsightly. small change, i changed to a switch case for the controlling of the animation & defined all of the circumstances.